Refactor back and landing button
