+++ src/components/Pages/Home/Browse.js
@@ -1,17 +1,17 @@
 import React from 'react';
-import SandButton from '../../buttons/Yellow';
-import { Title, Text, Section } from './homeStyles';
+import LandingButton from '../../buttons/LandingButton';
+import * as SC from './style';
 
 export default function Browse() {
-  return (
-    <Section>
-      <Title>Browse current radio workshops.</Title>
-      <Text>
-        Find specialize in audio workshops. Working with the workshop attendees on a topic with the end goal of producing a podcast or audio feature
-        to be aired on FM radio within the camp.
-      </Text>
+	return (
+		<SC.Section>
+			<SC.Title>Browse current radio workshops.</SC.Title>
+			<SC.Text>
+				Find specialize in audio workshops. Working with the workshop attendees on a topic with the end goal of producing a podcast or audio feature
+				to be aired on FM radio within the camp.
+			</SC.Text>
 
-      <SandButton innerText="Go To The Workshops" to="/workshops" />
-    </Section>
-  );
+			<LandingButton innerText="Go To The Workshops" to="/workshops" />
+		</SC.Section>
+	);
 }
