+++ src/components/Pages/Home/Home.js
@@ -6,8 +6,6 @@ import Contribute from './contribute';
 import FAQs from './faqs';
 
 export default function Home() {
-	//   const [currentTags, setCurrentTags] = React.useState([]);
-	// console.log('Home -> currentTags', currentTags);
 
 	return (
 		<>
