+++ src/components/Pages/Home/contribute.js
@@ -1,11 +1,13 @@
 import React from 'react';
-import { Title, Text, Section } from './homeStyles';
+import * as SC from './style';
 
 export default function Contribute() {
-  return (
-    <Section>
-      <Title>How to contribute</Title>
-      <Text>Lorem ipsum dolor sit amet, consectetur adipiscing elit ut aliquam, purus sit amet luctus venenatis, lectus magna fringilla urna.</Text>
-    </Section>
-  );
+	return (
+		<SC.Section>
+			<SC.Title>How to contribute</SC.Title>
+			<SC.Text>
+				Lorem ipsum dolor sit amet, consectetur adipiscing elit ut aliquam, purus sit amet luctus venenatis, lectus magna fringilla urna.
+			</SC.Text>
+		</SC.Section>
+	);
 }
