+++ src/styles/index.js
@@ -1,2 +1,2 @@
-export * as vars from './variables';
-export * as componentsStyles from './componentsStyles';
+// export * as vars from './variables';
+// export * as componentsStyles from './componentsStyles';
