+++ src/styles/variables.js
@@ -3,3 +3,5 @@ export const olive = '#b1b493';
 export const ocean = '#4f8a8b';
 export const night = '#07031a';
 export const offWhite = '#fbf9fa';
+export const medButtonHeight = '42px'
+export const smallButtonHeight = '38px'
